refactor: 프론트 구조 변경
