+++ react-native-app/screens/businessCard/CreateBusinessCardScreen.js
